backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-jwt-extended==4.5.2
flask-sqlalchemy==3.0.5
bcrypt==4.0.1
python-dotenv==1.0.0
gunicorn==21.2.0 